requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile requirements.in
#
apscheduler==3.7.0
    # via -r requirements.in
cassandra-driver==3.25.0
    # via -r requirements.in
click==8.0.1
    # via geomet
colorama==0.4.4
    # via click
geomet==0.2.1.post1
    # via cassandra-driver
importlib-metadata==4.6.1
    # via click
numpy==1.21.0
    # via pandas
pandas==1.3.0
    # via -r requirements.in
python-dateutil==2.8.1
    # via pandas
pytz==2021.1
    # via
    #   -r requirements.in
    #   apscheduler
    #   pandas
    #   tzlocal
six==1.16.0
    # via
    #   apscheduler
    #   cassandra-driver
    #   geomet
    #   python-dateutil
typing-extensions==3.10.0.0
    # via importlib-metadata
tzlocal==2.1
    # via apscheduler
zipp==3.5.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
